requirements/base.txt
# Stable Django
django==1.9.3

# mysql
mysqlclient==1.3.7

# Models
django-model-utils==2.3.1
django-dirtyfields==0.8.1

# Configuration
django-environ==0.4.0
django-secure==1.0.1

# Time zones support
python-dateutil==2.4.2
pytz==2015.6

# Redis support
django-redis==4.2.0
redis==2.10.3

# celery for async tasks
celery==3.1.20

# DRF
djangorestframework==3.3.2
cryptography==1.2.3
djangorestframework-jwt==1.7.2

pyjwkest==1.1.0
django-cors-headers==1.1.0

django-rest-swagger==0.3.5

freezegun== 0.3.6
unittest-xml-reporting==1.14.0
raven==5.11.1

# WSGI Handler
# ------------------------------------------------
gevent==1.1rc4
eventlet==0.18.2
gunicorn==19.3.0

# PL
-e git+ssh://git@github.com/presencelearning/python-pl.git@3db4bc80acf45b1d23e711bfd0e049560b518a98#egg=pl

# Static and Media Storage
# ------------------------------------------------
# boto==2.38.0
# django-storages-redux==1.3